Minor Civ Ratio in Doubles
Switches the Minor Civ Ratio calculating functions to return doubles
instead of floats.

+++ CvGameCoreDLL_Expansion2/CvGame.h
@@ -145,8 +145,8 @@ public:
 	int getNumSequentialHumans(PlayerTypes ignorePlayer = NO_PLAYER);
 
 #ifdef AUI_MINOR_CIV_RATIO
-	float getCurrentMinorCivRatio();
-	float getCurrentMinorCivDeviation();
+	double getCurrentMinorCivRatio();
+	double getCurrentMinorCivDeviation();
 #endif // AUI_MINOR_CIV_RATIO
 
 	int getGameTurn();
